backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.3
motor==3.3.2
beanie==1.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
apscheduler==3.10.4
openai==1.3.5
google-generativeai==0.3.1
python-multipart==0.0.6
dnspython==2.4.2
